✨ MBK-45: add shell command execution functionality
